third_party/blink/web_tests/css-parser/font-size-webkit-xxx-large.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/property-parsing-test.js"></script>
<script>
// Verifies that font-size property's -webkit-xxx-large value is properly parsed

assert_valid_value("fontSize", "-webkit-xxx-large");
</script>
